Fix attribute parsing when node contains LabeledText
Node attributes are now read based on last occurence of ']'-character
instead of its string representations length. Nodes with
LabeledText values containing the ']'-character are still likely to
have a deformed value for their last attribute, depending on the
location of the character. This has been added to the TODO list.