Fix Finder recursion when chained selectors are used and no root is set
Using chained selectors when looking for nodes from every window caused
a StackOverflowException because the list of remaining root nodes was
not updated properly. The list is used to determine whether the
recursion should continue or not, so the method was calling itself until
the StackOverflowException was thrown.

Finder now holds a reference to the original root node of the lookup so
the list of remaining root nodes actually updates. This commit also
fixes an UnsupportedOperationException in Find All which occured when
searching nodes from multiple windows. Exception happened because the
find results were added to the map recursively, and some of the results
were located in unmodifiable collections. Now a single Map gets passed
as a parameter for the whole duration of the process, and all found
nodes are added to this instance.