Move the lookup logic into a separate class
Moved lookup logic from Finder to the new FindOperation class. Finder
manages the state of the whole lookup process, like remaining roots and
recursion, while FindOperation performs a single lookup for a single root.

Find and Find All both use the same executeLookup method for the lookup
and cast the results respectively into either a Node or a Set. The type
of lookup (first/all) is given as an argument in FindOperations constructor.

Methods for getting the FindPrefix and removing it from the query string
were moved to QueryParser.